+++ src-docs/src/views/color_picker/props.tsx
@@ -1,0 +1,7 @@
+import React, { FunctionComponent } from 'react';
+
+import { EuiColorPalettePickerPaletteProps } from '../../../../src/components/color_picker/color_palette_picker';
+
+export const EuiColorPalettePickerPalette: FunctionComponent<
+  EuiColorPalettePickerPaletteProps
+> = () => <div />;