+++ src/components/color_picker/color_palette_picker/index.ts
@@ -17,11 +17,6 @@
  * under the License.
  */
 
-export { EuiColorPicker, EuiColorPickerProps } from './color_picker';
-export { EuiColorPickerSwatch } from './color_picker_swatch';
-export { EuiHue } from './hue';
-export { EuiSaturation } from './saturation';
-export { EuiColorStops } from './color_stops';
 export {
   EuiColorPalettePicker,
   EuiColorPalettePickerProps,